Refactor

+++ TextDifficultyDeterminer.Domain/Models/TextScores.cs
@@ -5,7 +5,7 @@ public class TextScores
     public string Name {get; set;}
     public int RealisticReadingThreshold {get; set;}
     public int ExtendedReadingThreshold {get; set;} 
-    public int WordCount {get; set;}
+    public long WordCount {get; set;}
     public int UniqueWords {get; set;}
     public Statistic Sentences {get; set;}
